+++ client/src/views/Home.js
@@ -1,0 +1,8 @@
+import React , {useState, useEffect} from "react";
+import {Link} from "react-router-dom";
+
+function Home(props) {
+    return <div>Home</div>
+}
+
+export default Home;
