+++ client/src/views/NotFound.js
@@ -1,0 +1,5 @@
+const NotFound = () => {
+    return "Page Not Found"
+}
+
+export default NotFound;
